separating app logic from roster generation task
