+++ CREDITS.md
@@ -1,0 +1,13 @@
+# Credits, Notes, and References
+
+   + http://code.tutsplus.com/tutorials/ruby-for-newbies-working-with-directories-and-files--net-18810
+   + http://stackoverflow.com/a/8660782/670433
+   + http://stackoverflow.com/a/13399580/670433
+   + http://stackoverflow.com/a/7727416/670433
+   + http://ruby-doc.org/core-2.2.0/Dir.html
+   + https://github.com/debate-watch/twenty_sixteen/blob/9ea59464df64cabc1236991bf12b3ff6a9162a3b/lib/twenty_sixteen/candidate.rb
+   + https://github.com/s2t2/branford_station/blob/1f7fc32f2788e95b0748c3e96645b27d792e3f79/app/workers/google_transit_data_feed_extractor.rb
+   + https://github.com/s2t2/branford_station/blob/1f7fc32f2788e95b0748c3e96645b27d792e3f79/spec/workers/google_transit_data_feed_extractor_spec.rb
+   + http://www.rubydoc.info/github/sparklemotion/nokogiri/Nokogiri/XML/Element
+   + http://stackoverflow.com/questions/4822422/output-array-to-csv-in-ruby
+   + http://stackoverflow.com/a/15912856/670433
